requirements.txt
flask
PyYAML
web.py==0.40-dev1
pymongo
PyMySQL
pytest